update 02.md Notes: useState & lazy initialization
useState takes 2 forms for the paramater passed in (function overloading)

    - send in a value (including from the result of an invoked function).
    - send in a function that it should invoke.

In both cases, the parameter is used only used the first time the component
renders, and ignored for all subsequent re-renders.

When passed a function, the hook invokes it internally. In contrast, if an
_invoked_ function is in the argument list, the is invoked by
the Componentat *every* render, to get that expensive function's result,
_before_ useState hook is called. This is normal 'expression evaluation'
from a function call's argument list. Normal behaviour. But it has
can have expensive implications: It means that the function is during
all re-renders, even though the _result_ of
that function call is ignored by `setState`! This causes unnecessary
bottleneck, slowness, if the function is computationally expensive.
(eg reading localStorage)

The practice of passing in a (non-invoked) function (usually in the form
of an anon function: `setState( ()=> {...; return expensiveInitialValue})`)
is known as _lazy initialization_ because the function is invoked only
when it is needed, ie at first render.

----
  Changes to be committed:
 	modified:   src/exercise/02.md