prep 02-extra-4 starting code and instructions
using JSON.parse and JSON.stringify so localStorage can save any data
type (ie objects and arrays) as opposed to the default, which can only
store strings and values (numbers, undefined, null) coerced into strings
specifically, we want to be able to store arrays and objects in the
localStorage, and read them back into the program as their orig data type.
I wonder if that would include functions, symbols, etc. I would think not,
but.. what if I want to store a function?

Also, it seems to me that I want to distinguish between the value
undefined/null/NaN/... and the strings 'undefined', 'null', 'NaN', 'Infinity'...

----
 On branch sh
 Changes to be committed:
	modified:   src/exercise/02.extra-4.js