fin 02-extra-4 basic implementation any data type
(incomplete, and comments must be deleted.
a basic implementation for some data types has been implemented
)

Cannot store runctions, symbols, undefined, bigInt.
Set gets stored as an Array. Etc. Not robust

I did not bother to create a handler with say a Submit button to
    allow the data bo change, and see if localStore would correctly
    update due to data changes.

I am not going to work any further on this. I imagine the exercise was
    not for us to properly store every data type in localStore.
    Indeed, it is not appropriate to be used as longterm persistant
    storage anyway. And JSON is an improper tool. It is far too limited
    to be used in this way.

TODO: Delete hundred lines of commentary!

----
 On branch sh
 Changes to be committed:
	modified:   src/exercise/02.extra-4.js