04.extra-3: refactor logic TO Game from Board
As per the solution file, I should nono longer keep ALL logic in the Board.
It should move UP to the Game component. Then strip the Board component
down to just displaying the board (and responding to clicks, that is due
to the callbackk being passed down from the parent Game component.)
 Everything else is now moved to Game.

    // Most logic was moved from Board component TO Game component
    //  Board only prints the gameboard and recieves click events
    //  for the squares. But the square onclick handler is in parent
    //  Game component. It is passed down through props.

    //  Game component now has squares, selectSquare, and everything
    //    to do with history.
    //  It also now renders the reset button, status, and the history buttons.

----
 On branch sh
	modified:   src/exercise/04.extra-3.js