BUG FIX: 05 vanilla tilt. METHOD, not LIBRARY call
The destroy() is a METHOD added to an element during init.
VanillaTilt.init() is the only function call.
VanillaTilt is the library script.
vanillaTilt.destroy is a METHOD added to the INSTANCE
	(notice the first letter is NOT capitalized)

This explains why init(element) -- the element is passed in.
VS the other methods do NOT take arguments, but are called ON the element:
element.vanillaTilt.destroy()		// vanillaTilt is a method ON element
																//	(it was created/added via below:)
VanillaTilt.init(element)				// creates an instance of VanillaTilt
																// script on the element
																//  (or dom node or nodeList array)
																// adds vanillaTilt.destroy,
																//	vaniallaTilt.reset, and
																//  vanillaTilt.?something else no remember?

----
 On branch sh
	modified:   src/exercise/05.js