06.extra-3 more COMMENTS setState in asynch funcs
add tons of comments and console.logs
updated minor bugs RE: logging and implications of where logs were.

special attention paid to RE: ordering of useState and console.log output
This has implications much beyond console.logs.
TLDR; order matters inside asynch func.
TBF, I should not be printing out a variable that is not included as
	a dependancy. It should only be printed OUTSIDE the useEffect function.
	(If I added it to the dependacy list, though, it would infinite-loop)

This is a lesson as to why the dependacy array is so important!

It also provides an interesting illustration of:
1) setState being both asynch, but also gauranteed for useEffect
2) Lifecycle order
3) closures inside useEffect, combined with Lifecycle order
5) the JS event loop, perhaps
4) Order matters inside asynch functions. // maybe that is not in play here.
		// my comments suggest it is, but I may have been mistaken. I now think
		// those other factors may account for everything.

Remember, I am keeping comments and console.logs because this file is
	a teaching document, and a record of my learning process.

----
 On branch sh
	modified:   src/exercise/06.extra-3.js