WIP 06.extra-4 broke during refactor
WIP refactoring to use an ErrorFallback function for the UI for an
	errorBoundary component.
But something is wrong.
If first enter/render an invalid pokemon
Then enter a valid pokemon
the errorBoundary UI gets rendered, and console.log shows it has the
	state data from the previously invalid pokemon

It also shows that there is an error in the ErrorBoundary itself,
	which is always troublesome & to be avoided always, since it cannot
	catch itself, and the whole white screen of death thing we were
	trying to avoid, cannot be averted.

Note: if first do a valid pokemon name, it does work.
And if follow that by an invalid name, that invalid UI also works.
BUT after that, it does not recover! in fact, if follow up by another
	but different invalid name, it still rerenders with the first invalid
	name. Basically, it is not resetting properly.

----
 On branch sh
	modified:   src/exercise/06.extra-4.js