PR suggestion 06.md, RE 06.extra-4 errorBoundary
TODO: create a PR for adding below line in 06.md (main branch, Dodds repo)
	at the end fo the section for 06.extra-4, so nobody else wastes a day
	thinking they did the exercise incorrectly or they had a bug.

Don't worry that your app does not retain functionality after your error
boundary component handles the error. We will cover that in the next extra
credit. Here we are just rewriting the UI so we do not crash it to a white
screen of death.

----
 On branch sh
	modified:   src/exercise/06.md