Syn 06.extra-4 errorBoundary RE: PR suggestion
TODO: create a PR for adding below line in 06.md (main branch, Dodds repo)
	at the end of the section for 06.extra-4, so nobody else wastes a day
	thinking they did the exercise incorrectly or they had a bug.

Don't worry that your app does not retain functionality after your error
boundary component handles an error. We will cover that in the next extra
credit. Here the error boundary is just replacing a broken UI so the app
does not crash to a white screen of death.

----
 On branch sh
	modified:   src/exercise/06.md